Implemented all Services from the model package

+++ src/main/java/com/marcosoft/storageSoftware/repository/CategoryRepository.java
@@ -3,7 +3,8 @@ package com.marcosoft.storageSoftware.repository;
 import com.marcosoft.storageSoftware.model.Category;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoryRepository extends CrudRepository<Category, Integer> {
+public interface CategoryRepository extends CrudRepository<Category, Long> {
 
-  Category findByCategoryName(String categoryName);
+
+    Category findByCategoryName(String categoryName);
 }
